requirements.txt
asgiref==3.5.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.8.0
bcc==0.18.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
channels==3.0.5
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
cloudinary==1.34.0
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==41.0.2
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
Django==3.2.12
django-cleanup==5.2.0
django-crispy-forms==2.0
django-multiupload==0.6.1
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
env==0.1.0
environ==1.0
Flask==2.0.1
hidpidaemon==18.4.6
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==22.10.0
itsdangerous==2.1.0
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
MarkupSafe==2.0.1
more-itertools==8.10.0
natsort==8.0.2
netaddr==0.8.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
Pillow==9.0.1
pop-transition==1.1.2
protobuf==3.12.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.5.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2021.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
SecretStorage==3.3.1
service-identity==23.1.0
sessioninstaller==0.0.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.2.4
systemd-python==234
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==2.0.2
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.0
